tracking: do not update last_message on msg pin
When a message is pinned, this will fire raw_edit_message, even though
the user did not actually edit the message (and the pinner is unknown).
If we get a message edit with no edited_timestamp, or with edit
timestamp in the past, ignore it as it is probably a pin.