refactor plugin/extension loading
Plugin loading now is based on module names - these modules must be able
to be imported, so must be on the path etc. For old style directory
loading use the "*" e.g. "dango.plugins.*", or you can load just
individual modules. Note omitting the "*" will result in not importing
submodules, just the __init__.py in there.

Also move all plugins to dango/, instead of being on the root namespace.

This is going to break existing configs... :)