fix storing last_seen in UTC
As it turns out...

If you call datetime.utcnow().timestamp(), the timestamp returned does
*not* correspond with current utc time... And is dependent on local time
zone. That is to say, datetime.utcnow().timestamp() will return
different values in different timezones at the same time. The proper way
to get a utc timestamp is, instead, if given a naive timestamp in UTC,
is to first .replace(tzinfo=timezone.utc), then call timestamp(), and
then call utcfromtimestamp() when converting to timestamp to get in in
UTC again....